Refactoring: Properly use Services and DTO
